==> pre_processing/Part1_search_labkey.ipynb <==
Import modules

import numpy as np
import functools
import labkey
import pandas as pd

Helper function 

def labkey_to_df(sql_query, database, maxrows):
  """generate an pandas dataframe from labkey sql query

  Args:
    sql_query (str): an sql query as string.
    dr (str): GEL datarelease version
    """
  ver = labkey.__version__

  if ver == '1.2.0' or ver == '1.4.0' or ver == '1.4.1':
    server_context = labkey.utils.create_server_context(
      domain= "labkey-embassy.gel.zone",
      container_path = database,
      context_path = "labkey",
      use_ssl = True
    )

    results =  labkey.query.execute_sql(
      server_context,
      schema_name="lists",
      sql=sql_query,
      max_rows=maxrows
    )

  elif ver == '2.4.0':
    from labkey.api_wrapper import APIWrapper

    labkey_server = "labkey-embassy.gel.zone"
    project_name = database  # Project folder name
    contextPath = "labkey"
    schema = 'lists'
    api = APIWrapper(
    labkey_server, 
    project_name, 
      contextPath, 
      use_ssl=True)

    results = api.query.execute_sql(sql=sql_query, schema_name=schema, max_rows = maxrows)

  return(pd.DataFrame(results['rows']))

100kGP

version = "/main-programme/main-programme_v18_2023-12-21"

Case cohort

disease = "INSERT DISEASE"

recruited_sql = (f'''
    SELECT participant_id, 
        normalised_specific_disease, 
        normalised_disease_sub_group, 
        normalised_disease_group
    FROM rare_diseases_participant_disease
    WHERE normalised_specific_disease = '{disease}'
    ''')

recruited_query = labkey_to_df(recruited_sql, version, 10000)
recruited_query

recruited_participants = recruited_query['participant_id'].tolist()

participants = list(dict.fromkeys(recruited_participants))
len(participants)

solved = "yes"

gmc_sql = (f'''
    SELECT participant_id, case_solved_family
    FROM gmc_exit_questionnaire
    WHERE participant_id in {*participants,}
    AND case_solved_family != '{solved}'
    ''')

gmc_query = labkey_to_df(gmc_sql, version, 1000)
gmc_query

dd_sql = (f'''
    SELECT participant_id
    FROM submitted_diagnostic_discovery
    WHERE participant_id IN {*gmc_query['participant_id'].tolist(),}
    ''')

dd_query = labkey_to_df(dd_sql, version, 1000)
dd_query

cohort = [i for i in gmc_query['participant_id'].tolist() if i not in dd_query['participant_id'].tolist()]
len(cohort)

ICD 10 cases

Control cohort

not_disease = "Skeletal disorders"

recruited_not_sql = (f'''
    SELECT participant_id, 
        normalised_specific_disease, 
        normalised_disease_sub_group, 
        normalised_disease_group
    FROM rare_diseases_participant_disease
    WHERE normalised_disease_group = '{not_disease}'
    ''')

recruited_not_query = labkey_to_df(recruited_not_sql, version, 100000)
recruited_not_list = recruited_not_query['participant_id'].tolist()
len(recruited_not_list)

icd_not_codes = ['Q30.0','Q77.3','Q75.0','78.3','Q78.0']

hes_tables = ["apc", "op", "ae"]

icd_not_concat = pd.DataFrame()

diag_not_statement = "%' OR diag_all LIKE '%".join(icd_not_codes)

for hes_table in hes_tables:
    sqlstr = (
        f'''
        SELECT participant_id, diag_all
        FROM hes_{hes_table}
        WHERE diag_all LIKE '%{diag_not_statement}%'
        ''')
    icd_query = labkey_to_df(sqlstr, version, 1000000)
    if not icd_query.empty:
        icd_not_concat = pd.concat([icd_query, icd_not_concat])

icd_not_list = icd_not_concat['participant_id'].tolist()
len(icd_not_list)

not_list = recruited_not_list + icd_not_list 
not_list = list(dict.fromkeys(not_list))
len(not_list)

participant_sql = 'SELECT participant_id FROM participant'
participant_list = labkey_to_df(participant_sql, version, 1000000)['participant_id'].tolist()
control = [i for i in participant_list if i not in not_list]
len(control)

Match demographics

from statistics import mean

case_sql = (f'''
    SELECT participant_id, yob, 
    participant_phenotyped_sex, participant_karyotyped_sex
    FROM participant_summary
    WHERE participant_id IN {*cohort,}
    ''')
case_query = labkey_to_df(case_sql, version, 10000).drop_duplicates()
case_yob_mean = mean(case_query['yob'].tolist())

control_sql = (f'''
    SELECT participant_id, yob, 
    participant_phenotyped_sex, participant_karyotyped_sex
    FROM participant_summary
    WHERE participant_id IN {*control,}
    ''')
control_query = labkey_to_df(control_sql, version, 100000).drop_duplicates()
control_yob_mean = mean(control_query['yob'].tolist())

print("case mean = ", case_yob_mean, 
      "\ncontrol mean = ", control_yob_mean)

print("case XX = ", case_query['participant_karyotyped_sex'].value_counts()['XX']/len(case_query),
     "\ncase Female = ", case_query['participant_phenotyped_sex'].value_counts()['Female']/len(case_query),
     "\ncase XY = ", case_query['participant_karyotyped_sex'].value_counts()['XY']/len(case_query),
     "\ncase Male = ", case_query['participant_phenotyped_sex'].value_counts()['Male']/len(case_query),
     "\ncontrol XX = ", control_query['participant_karyotyped_sex'].value_counts()['XX']/len(control_query),
     "\ncontrol Female = ", control_query['participant_phenotyped_sex'].value_counts()['Female']/len(control_query),
     "\ncontrol XY = ", control_query['participant_karyotyped_sex'].value_counts()['XY']/len(control_query),
     "\ncontrol Male = ", control_query['participant_phenotyped_sex'].value_counts()['Male']/len(control_query))

eth = "European"

case_ethnicity_sql = (f'''
    SELECT participant_id, genetically_inferred_ancestry_thr
    FROM participant_summary
    WHERE genetically_inferred_ancestry_thr = '{eth}'
    AND participant_id IN {*cohort,}
    ''')
case_ethnicity_query = labkey_to_df(case_ethnicity_sql, version, 10000).drop_duplicates()
filtered_case = case_ethnicity_query['participant_id'].tolist()

control_ethnicity_sql = (f'''
    SELECT participant_id, genetically_inferred_ancestry_thr
    FROM participant_summary
    WHERE genetically_inferred_ancestry_thr = '{eth}'
    AND participant_id IN {*control,}
    ''')
control_ethnicity_query = labkey_to_df(control_ethnicity_sql, version, 100000).drop_duplicates()
filtered_control = control_ethnicity_query['participant_id'].tolist()

General inclusion critera

case_df = pd.DataFrame(filtered_case, columns = ['participant_id'])
case_df['case'] = "case"
control_df = pd.DataFrame(filtered_control, columns = ['participant_id'])
control_df['case'] = "control"

case_control_table = pd.concat([case_df, control_df]).drop_duplicates()
case_control_table

sex_sql = (f'''
    SELECT participant_id, participant_phenotyped_sex, participant_karyotyped_sex
    FROM participant_summary
    WHERE participant_id IN {*case_control_table['participant_id'].tolist(),} 
    AND (participant_karyotyped_sex = 'XX'
    OR participant_karyotyped_sex = 'XY')
    ''')
sex_query = labkey_to_df(sex_sql, version, 100000).drop_duplicates()

sex_query = sex_query.drop(sex_query[ (sex_query['participant_phenotyped_sex'] == 'Male') & (sex_query['participant_karyotyped_sex'] == 'XX') ].index)
sex_query = sex_query.drop(sex_query[ (sex_query['participant_phenotyped_sex'] == 'Female') & (sex_query['participant_karyotyped_sex'] == 'XY') ].index)

case_control_table = case_control_table.loc[
    case_control_table['participant_id'].isin
    (sex_query['participant_id'].tolist())]
case_control_table

aggregate_sql = (f'''
    SELECT participant_id, platekey
    FROM aggregate_gvcf_sample_stats
    WHERE participant_id IN {*case_control_table['participant_id'].tolist(),}
    AND sample_source = 'BLOOD'
    AND sample_preparation_method = 'EDTA'
    AND sample_library_type = 'TruSeq PCR-Free High Throughput'
    ''')
aggregate_query = labkey_to_df(aggregate_sql, version, 100000)
case_control_table = aggregate_query.merge(case_control_table).drop_duplicates()
case_control_table

relatedness_table = pd.read_csv('/gel_data_resources/main_programme/aggregation/aggregate_gVCF_strelka/aggV2/additional_data/PCs_relatedness/relatedness/GEL_aggV2_MAF5_mp10_0.0442.kin0',sep = '\t')
relatedness_table

mz_twins_table = relatedness_table[ (relatedness_table['KINSHIP'] > 0.354) ]
mz_twins_table

twins_list = mz_twins_table[['IID1', 'IID2']].values.tolist()

case_control_list = case_control_table['platekey'].tolist()

for twins in twins_list:
    if twins[0] in case_control_list:
        if (case_control_table.query('platekey == @twins[0]')['case']).values == 'case':
            case_control_table = case_control_table.drop(case_control_table[case_control_table['platekey'] == twins[1]].index)
            print (twins[1])
    elif twins[1] in case_control_list:
        if (case_control_table.query('platekey == @twins[1]')['case']).values == 'case':
            case_control_table = case_control_table.drop(case_control_table[case_control_table['platekey'] == twins[0]].index)
            print(twins[0])
case_control_table

Filepaths

filetype = "Genomic VCF"

path_sql = (f'''
    SELECT participant_id, platekey, filename, file_path
    FROM genome_file_paths_and_types
    WHERE participant_id IN {*case_control_table['participant_id'].tolist(),}
    AND file_sub_type = '{filetype}'
''')

path_query = labkey_to_df(path_sql, version, 100000)
case_control_paths = path_query.merge(case_control_table)
case_control_paths

count_control = (case_control_paths['case'] == 'control').sum()
count_case = (case_control_paths['case'] == 'case').sum()
print(f'case - {count_case}')
print(f'control - {count_control}')

case_paths = case_control_paths[case_control_paths['case'] == 'case']
control_paths = case_control_paths[case_control_paths['case'] == 'control']
case_paths.to_csv('Part1_case_paths_output.csv', index=False)
control_paths.to_csv('Part1_control_paths_output.csv', index=False)
case_sample_size = pd.DataFrame([count_case])
case_sample_size.to_csv('case_sample_size.csv', index=False, header=False)
